refactor: extract Country type from countries.server
