+++ src/app/admin/gigs/_lib/countries.server.ts
@@ -2,10 +2,9 @@ import 'server-only';
 
 import { apiPublicRequest } from '@/lib/api-public';
 import { parseCountries } from '@/lib/api-boundary-schemas';
+import type { Country } from '@/app/admin/gigs/_lib/countries.types';
 
-export interface Country {
-  iso: string;
-}
+export type { Country };
 
 export async function getCountries(): Promise<Country[]> {
   const raw = await apiPublicRequest<unknown>('/v1/location/countries', 'GET', undefined, {
