+++ src/app/admin/gigs/_lib/countries.types.ts
@@ -1,0 +1,3 @@
+export interface Country {
+  readonly iso: string;
+}
